core: add proper fix for #2
The reason of the issue is memory addressing and different processing
of duplicate vertices. In debug, uninitialized. heap memory is filled with
0xcdcdcdcd in case of VS compiler which breaks comparison with NULL.

In release for Win32, it is not reproducible. However, in release x64 it is
reproducible as well.

Solution is to check type of vertex, as duplicates have UNDEADVERTEX
type.